feat: total info component create
