+++ client/src/components/profile/TypingStatistic/index.tsx
@@ -33,22 +33,30 @@ export default function TypingStatistic() {
   const [selectStatistic, setSelectStatistic] = useState(TypingStatisticEnum.ACCURACY);
 
   return (
-    <Box w='727px'>
+    <>
       <Flex padding='16px 5px' alignItems='center' justifyContent='space-between'>
         <Text as='h2' textStyle='text/hd4'>
           타자 통계
         </Text>
 
-        <Text textStyle='text/medium' colorScheme='white' color='gray.dark' h='26px' w='fit-content'>
+        <Text cursor='pointer' textStyle='text/medium' colorScheme='white' color='gray.dark' h='26px' w='fit-content'>
           전체보기
         </Text>
       </Flex>
-      <Box border='0.6px solid' borderColor='gray/main' bg='white/light' borderRadius='10px' padding='21px 28px'>
+      <Box
+        border='0.6px solid'
+        borderColor='gray.main'
+        bg='white.light'
+        borderRadius='10px'
+        padding='21px 28px'
+        h='353px'
+      >
         <Flex gap='16px'>
           <Text
             color={selectStatistic === TypingStatisticEnum.ACCURACY ? 'black.main' : 'gray.dark'}
             textStyle='text/medium'
             onClick={() => setSelectStatistic(TypingStatisticEnum.ACCURACY)}
+            cursor='pointer'
           >
             {TypingStatisticEnum.ACCURACY}
           </Text>
@@ -56,6 +64,7 @@ export default function TypingStatistic() {
             color={selectStatistic === TypingStatisticEnum.SPEED ? 'black.main' : 'gray.dark'}
             textStyle='text/medium'
             onClick={() => setSelectStatistic(TypingStatisticEnum.SPEED)}
+            cursor='pointer'
           >
             {TypingStatisticEnum.SPEED}
           </Text>
@@ -63,6 +72,6 @@ export default function TypingStatistic() {
 
         <LineChart chartTitle='타자 통계' chartData={CharTestData} />
       </Box>
-    </Box>
+    </>
   );
 }
